Condense this.rules function
The this.rules function was condensed in a similar manner suggested by the java implementation comment above the this.rules function.

The previous implementation assigned the binary values backwards, i.e. 7 == index0, 6 === index1, etc.

This condensed version assigns them in order. 

I wasn't sure if this was intentional, it appears the ruleset is randomized so from my understanding this won't make a difference.